OLD/spike/requirements.txt
Flask==3.1.2
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.153.0
Werkzeug>=3.1.2
